+++ core/info/logging_handler.py
@@ -2,6 +2,8 @@
 
 """Following script can output stdout and stderr to Qt text browser."""
 
+from __future__ import annotations
+
 __author__ = "Yuan Chang"
 __copyright__ = "Copyright (C) 2016-2019"
 __license__ = "AGPL"
@@ -9,16 +11,25 @@ __email__ = "pyslvs@gmail.com"
 
 from typing import Optional
 import sys
-import logging
+import os
+from logging import (
+    DEBUG,
+    INFO,
+    basicConfig,
+    getLogger,
+    Handler,
+    StreamHandler,
+)
+from .info import ARGUMENTS
 from core.QtModules import QObject, Signal
 
 
-class _QtHandler(logging.Handler):
+class _QtHandler(Handler):
 
     """Logging handle."""
 
     def __init__(self):
-        logging.Handler.__init__(self)
+        super(_QtHandler, self).__init__()
 
     def emit(self, record: str):
         """Output to the other side."""
@@ -27,43 +38,48 @@ class _QtHandler(logging.Handler):
             XStream.stdout().write(record + '\n')
 
 
-_logger = logging.getLogger(__name__)
-_handler = _QtHandler()
-_handler.setFormatter(logging.Formatter("%(asctime)s | %(message)s"))
-_logger.addHandler(_handler)
-
 _SYS_STDOUT = sys.stdout
 _SYS_STDERR = sys.stderr
+_log_file_name = "./pyslvs.log"
+
+basicConfig(
+    level=DEBUG if ARGUMENTS.debug_mode else INFO,
+    filename=_log_file_name,
+    format="[%(asctime)s] [%(funcName)s]:%(levelname)s: %(message)s",
+)
+logger = getLogger()
+logger.addHandler(_QtHandler())
+logger.setLevel(DEBUG)
+_std_handler = StreamHandler(_SYS_STDOUT)
+logger.addHandler(_std_handler)
 
 
 class XStream(QObject):
 
     """Stream object to imitate Python output."""
 
-    _stdout: Optional['XStream'] = None
-    _stderr: Optional['XStream'] = None
-    messageWritten = Signal(str)
+    _stdout: Optional[XStream] = None
+    _stderr: Optional[XStream] = None
+    message_written = Signal(str)
+
+    @staticmethod
+    def flush():
+        """Remove log file if exit."""
+        os.remove(_log_file_name)
 
     def write(self, msg: str):
         """Output the message."""
         if not self.signalsBlocked():
-            self.messageWritten.emit(msg)
+            self.message_written.emit(msg)
 
     @staticmethod
-    def stdout() -> 'XStream':
+    def stdout() -> XStream:
         """Replace stdout."""
         if not XStream._stdout:
             XStream._stdout = XStream()
             sys.stdout = XStream._stdout
+            logger.removeHandler(_std_handler)
         return XStream._stdout
-
-    @staticmethod
-    def stderr() -> 'XStream':
-        """Replace stderr."""
-        if not XStream._stderr:
-            XStream._stderr = XStream()
-            sys.stderr = XStream._stderr
-        return XStream._stderr
 
     @staticmethod
     def back():
@@ -72,3 +88,4 @@ class XStream(QObject):
         sys.stderr = _SYS_STDERR
         XStream._stdout = None
         XStream._stderr = None
+        logger.addHandler(_std_handler)